Two python codes added to interact conveniently with the main code.
coordinate is a wrapper that takes the same command-line arguments
as the original model B code in C and translates them into an input
file for the C++ code.

coherence.py allows you to generate plots like the ones shown in
the coordinationgames.pdf notes on the original model B code.
It plots the degree of coherence within and between agents.